+++ pysat/pb.py
@@ -23,6 +23,15 @@
     Module description
     ==================
 
+    .. note::
+
+        Functionality of this module is available only if the `PyPBLib`
+        package is installed, e.g. from PyPI:
+
+        .. code-block::
+
+            $ pip install pypblib
+
     This module provides access to the basic functionality of the `PyPBLib
     library <https://pypi.org/project/pypblib/>`__ developed by the `Logic
     Optimization Group <http://ulog.udl.cat/>`__ of the University of Lleida.
@@ -78,8 +87,14 @@
 #
 #==============================================================================
 import math
-from pypblib import pblib
 from pysat.formula import CNF
+
+# checking whether or not pypblib is available and working as expected
+pblib_present = True
+try:
+    from pypblib import pblib
+except ImportError:
+    pblib_present = False
 
 
 #
@@ -241,6 +256,8 @@ class PBEnc(object):
 
             :rtype: :class:`pysat.formula.CNF`
         """
+
+        assert pblib_present, 'Package \'pypblib\' is unavailable. Check your installation.'
 
         if encoding < 0 or encoding > 5:
             raise(NoSuchEncodingError(encoding))